requirements.txt
# Core dependencies
pandas==2.2.2
numpy==1.26.4
PyYAML==6.0.2
tqdm==4.66.1

# HuggingFace
transformers==4.41.2
datasets==2.19.1
tokenizers>=0.19,<0.20

# Training / runtime
torch==2.2.2

# Utilities
scikit-learn==1.4.2
nltk==3.8.1
streamlit==1.35.0

# Logging
rich==13.3.3